split out sha code for legibility
